Made Consolidated LocalQuotesService
